high-mem-func-benchmarks/image-processing/requirements.txt
Pillow
numpy~=1.26.4
